Move tests for properties & helpers into tests/unit/-private folder
