+++ tests/unit/-private/properties/alias-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import {
   create,
   clickable,
