+++ tests/unit/-private/properties/count-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import { create, count } from 'ember-cli-page-object';
 
 moduleForProperty('count', function(test) {
