+++ tests/unit/-private/properties/is-hidden-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import { create, isHidden } from 'ember-cli-page-object';
 
 moduleForProperty('isHidden', function(test) {
