+++ tests/unit/-private/properties/triggerable-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import { create, triggerable } from 'ember-cli-page-object';
 
 moduleForProperty('triggerable', function(test) {
