+++ tests/unit/-private/properties/value-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import { create, value } from 'ember-cli-page-object';
 
 moduleForProperty('value', function(test) {
