+++ tests/unit/-private/properties/visitable-test.js
@@ -1,4 +1,4 @@
-import { moduleForProperty } from '../helpers/properties';
+import { moduleForProperty } from '../../../helpers/properties';
 import { create, visitable } from 'ember-cli-page-object';
 
 moduleForProperty('visitable', { acceptanceOnly: true }, function(test) {
